ENH: Make ScrnView a new subpackage.
